+++ tests/custom-call/test.php
@@ -1,0 +1,6 @@
+<?php
+
+declare(strict_types=1);
+
+$app->t('custom-test');
+$app->t('custom-test2');