+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.0.1 under development
 
-- no changes in this release.
+- Enh #42: Add ability to change the calling method in translator extractor. 
 
 ## 2.0.0 February 17, 2023